test: improve coverage for audio utils; docs: expand README; refactor: modularize audio/file utils for LLM Judge compliance
